requirements.txt
duckduckgo_search
gnews
matplotlib
newspaper3k
numpy
pandas
requests
scikit-learn
statsmodels